MemStruct/Types: Renamed analysis.mem to core.types
